Added tests, overhauled plotting.geodesics; tox updates
