+++ docs/source/api/plotting/geodesics/core.rst
@@ -3,6 +3,6 @@ Core plotting module
 
 This module contains the advanced backend switching while plotting.
 
-.. automodule:: einsteinpy.plotting.geodesics.core
+.. automodule:: einsteinpy.plotting.geodesic.core
     :members:
     :show-inheritance:
