+++ docs/source/api/plotting/geodesics/interactive.rst
@@ -3,6 +3,6 @@ Interactive plotting module
 
 This module contains interactive plotting backend.
 
-.. automodule:: einsteinpy.plotting.geodesics.interactive
+.. automodule:: einsteinpy.plotting.geodesic.interactive
     :members:
     :show-inheritance:
