+++ docs/source/jupyter.rst
@@ -17,7 +17,6 @@ Jupyter notebooks
    :maxdepth: 1
    :caption: metric, bodies, geodesic, coordinates
    
-   /examples/Analyzing Earth using EinsteinPy!.ipynb
    /examples/Animations in EinsteinPy.ipynb
    /examples/Visualizing Event Horizon and Ergosphere (Singularities) of Kerr Metric or Black Hole.ipynb
    /examples/Visualizing Frame Dragging in Kerr Spacetime using EinsteinPy!.ipynb
