+++ src/einsteinpy/geodesic/timelike.py
@@ -49,7 +49,6 @@ class Timelike(Geodesic):
 
         """
         super().__init__(
-            self,
             position=position,
             momentum=momentum,
             a=a,
