+++ src/einsteinpy/plotting/__init__.py
@@ -1,11 +1,6 @@
 from einsteinpy.plotting.fractal import fractal
-from einsteinpy.plotting.geodesics.core import GeodesicPlotter
-from einsteinpy.plotting.geodesics.interactive import InteractiveGeodesicPlotter
-from einsteinpy.plotting.geodesics.null.core import NullGeodesicPlotter
-from einsteinpy.plotting.geodesics.null.interactive import (
-    InteractiveNullGeodesicPlotter,
-)
-from einsteinpy.plotting.geodesics.null.static import StaticNullGeodesicPlotter
-from einsteinpy.plotting.geodesics.static import StaticGeodesicPlotter
+from einsteinpy.plotting.geodesic.core import GeodesicPlotter
+from einsteinpy.plotting.geodesic.interactive import InteractiveGeodesicPlotter
+from einsteinpy.plotting.geodesic.static import StaticGeodesicPlotter
 from einsteinpy.plotting.hypersurface.core import HypersurfacePlotter
 from einsteinpy.plotting.rays.shadow import ShadowPlotter
