+++ src/einsteinpy/plotting/geodesic/__init__.py
@@ -1,0 +1,3 @@
+from einsteinpy.plotting.geodesic.core import GeodesicPlotter
+from einsteinpy.plotting.geodesic.interactive import InteractiveGeodesicPlotter
+from einsteinpy.plotting.geodesic.static import StaticGeodesicPlotter
