--- src/einsteinpy/plotting/geodesics/__init__.py
@@ -1,8 +1,0 @@
-from einsteinpy.plotting.geodesics.core import GeodesicPlotter
-from einsteinpy.plotting.geodesics.interactive import InteractiveGeodesicPlotter
-from einsteinpy.plotting.geodesics.null.core import NullGeodesicPlotter
-from einsteinpy.plotting.geodesics.null.interactive import (
-    InteractiveNullGeodesicPlotter,
-)
-from einsteinpy.plotting.geodesics.null.static import StaticNullGeodesicPlotter
-from einsteinpy.plotting.geodesics.static import StaticGeodesicPlotter
