+++ src/einsteinpy/symbolic/predefined/barriola_vilenkin.py
@@ -6,7 +6,7 @@ from einsteinpy.symbolic.metric import MetricTensor
 
 def BarriolaVilekin(c=constants.c, k=symbols("k")):
     """
-    Barriola-Vilekin monopol metric 
+    Barriola-Vilekin monopol metric
     Phys. Rev. Lett. 63, 341
     Manuel Barriola and Alexander Vilenkin
     Published 24 July 1989
