+++ src/einsteinpy/symbolic/predefined/minkowski.py
@@ -13,7 +13,7 @@ def MinkowskiCartesian(c=constants.c):
     c : ~sympy.core.basic.Basic or int or float
         Any value to assign to speed of light. Defaults to 'c'.
 
-    
+
     """
     coords = symbols("t x y z")
     metric = diag(-1, 1 / (c ** 2), 1 / (c ** 2), 1 / (c ** 2)).tolist()
