+++ src/einsteinpy/symbolic/predefined/vacuum_solutions.py
@@ -122,7 +122,7 @@ def ReissnerNordstorm(
     sch=symbols("r_s"),
     Q=symbols("Q"),
 ):
-    """"
+    """ "
     The Reissner–Nordström metric in spherical coordinates
     A static solution to the Einstein–Maxwell field equations,
     which corresponds to the gravitational field of a charged,
