+++ tests/test_examples.py
@@ -1,6 +1,7 @@
+from numpy.testing import assert_allclose
+
 from einsteinpy.examples import perihelion
 from einsteinpy.geodesic import Timelike
-from einsteinpy.metric import Schwarzschild
 
 
 def test_perihelion_attr():
@@ -9,7 +10,9 @@ def test_perihelion_attr():
 
     """
     p = perihelion()
+    L = p.trajectory[1][0, 5]
+
     assert isinstance(p, Timelike)
-    assert p.state.shape[0] == 8
-    assert p.trajectory.shape[1] == 8
-    assert isinstance(p.metric, Schwarzschild)
+    assert p.trajectory[1].shape[1] == 6
+    assert_allclose(p.trajectory[1][:, 4], 0, atol=1e-12, rtol=1e-12)
+    assert_allclose(p.trajectory[1][:, 5], L, atol=1e-12, rtol=1e-12)
